requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.1
anthropic==0.40.0
openai==1.58.1
gitpython==3.1.43
google-generativeai==0.8.6
